feat: relocate all packages under parent play4s including api registry
